_requirements.txt
asgiref
attrs
autobahn
Automat
cbor2
certifi
cffi
channels
channels-redis
charset-normalizer
constantly
cryptography
daphne
Django
django-cors-headers
django-extensions
django-filter
django-seed
djangorestframework
djangorestframework-simplejwt
drf-yasg
Faker
gunicorn
hyperlink
idna
Incremental
inflection
msgpack
packaging
psycopg2
py-ubjson
pyasn1
pyasn1_modules
pycparser
PyJWT
pyOpenSSL
python-dateutil
python-dotenv
pytz
PyYAML
redis
requests
service-identity
setuptools
six
sqlparse
toposort
Twisted
txaio
typing_extensions
tzdata
ujson
uritemplate
urllib3
zope.interface